# 8-buffer PreSieve (used to be 4-buffer), estimated 1%-4% speed improvement