--- a/src/PreSieve.cpp
+++ b/src/PreSieve.cpp
@@ -1,15 +1,19 @@
 ///
 /// @file   PreSieve.cpp
-/// @brief  Pre-sieve multiples of small primes <= 59 to speed up the
+/// @brief  Pre-sieve multiples of small primes < 100 to speed up the
 ///         sieve of Eratosthenes. The idea is to allocate several
 ///         arrays (buffers_) and remove the multiples of small primes
 ///         from them at initialization. Each buffer is assigned
 ///         different primes, for example:
 ///
-///         Buffer 0 removes multiplies of:  7, 19, 23, 29
-///         Buffer 1 removes multiplies of: 11, 13, 17, 37
-///         Buffer 2 removes multiplies of: 31, 47, 59
-///         Buffer 3 removes multiplies of: 41, 43, 53
+///         Buffer 0 removes multiplies of: {  7, 61, 73 }
+///         Buffer 1 removes multiplies of: { 11, 53, 59 }
+///         Buffer 2 removes multiplies of: { 13, 37, 71 }
+///         Buffer 3 removes multiplies of: { 17, 29, 67 }
+///         Buffer 4 removes multiplies of: { 19, 41, 43 }
+///         Buffer 5 removes multiplies of: { 23, 31, 47 }
+///         Buffer 6 removes multiplies of: { 79, 97 }
+///         Buffer 7 removes multiplies of: { 83, 89 }
 ///
 ///         Then whilst sieving, we perform a bitwise AND on the
 ///         buffers_ arrays and store the result in the sieve array.
@@ -143,7 +147,7 @@ const std::array<uint8_t, 7*11*13> buffer_7_11_13 =
   0xc7
 };
 
-/// Pre-sieve with the primes <= 97
+/// Pre-sieve with the primes < 100
 const std::array<std::vector<uint64_t>, 8> bufferPrimes =
 {{
   {  7, 61, 73 },  // 31 KiB
@@ -153,7 +157,7 @@ const std::array<std::vector<uint64_t>, 8> bufferPrimes =
   { 19, 41, 43 },  // 33 KiB
   { 23, 31, 47 },  // 33 KiB
   { 79, 97 },      //  8 KiB
-  { 83, 89 },      //  7 KiB
+  { 83, 89 }       //  7 KiB
 }};
 
 void andBuffers(const uint8_t* __restrict buf1,
@@ -248,14 +252,12 @@ void PreSieve::preSieve(uint8_t* sieve,
   else
     preSieveLarge(sieve, sieveSize, segmentLow);
 
-  // Pre-sieving removes the primes <= 59. We
+  // Pre-sieving removes the primes < 100. We
   // have to undo that work and reset these bits
   // to 1 (but 49 = 7 * 7 is not a prime).
   uint8_t bit49 = 1 << 4;
   uint8_t bit77 = 1 << 3;
   uint8_t bit91 = 1 << 7;
-  uint8_t bit119 = 1 << 6;
-  uint8_t bit121 = 1 << 7;
   size_t i = 0;
 
   if (segmentLow < 30)
@@ -264,8 +266,6 @@ void PreSieve::preSieve(uint8_t* sieve,
     sieve[i++] = 0xff ^ bit49;
   if (segmentLow < 90)
     sieve[i++] = 0xff ^ bit77 ^ bit91;
-  if (segmentLow < 120)
-    sieve[i++] = 0xff ^ bit119 ^ bit121;
 }
 
 /// Pre-sieve with the primes <= 13
@@ -296,7 +296,7 @@ void PreSieve::preSieveSmall(uint8_t* sieve,
   }
 }
 
-/// Pre-sieve with the primes <= 59
+/// Pre-sieve with the primes < 100
 void PreSieve::preSieveLarge(uint8_t* sieve,
                              uint64_t sieveSize,
                              uint64_t segmentLow) const