I've been trying to squeeze a bit more performance of out the multi-buffer PreSieve idea, and at the same time reduce the total memory taken by these buffers. Let me know what you think.

**Single-core performance**
`primesieve 1e11 -t1`
| Compiler | -march | Original (seconds) | Modified (seconds) | Improvement | Original (Giga-cycles) | Modified (Giga-cycles) | Improvement
| --- | --- | --- | --- | --- | --- | --- | --- |
clang | skylake | 14.680 | 14.262 | 2.85% | 60.729 | 59.219 | 2.49%
clang | x86-64 | 14.811 | 14.370 | 2.98% | 61.064 | 59.400 | 2.72%
gcc | skylake | 14.800 | 14.588 | 1.43% | 61.541 | 60.700 | 1.37%
gcc | x86-64 | 15.027 | 14.436 | 3.93% | 62.720 | 60.142 | 4.11%

**Multi-core performance**
`primesieve 1e11`
| Compiler | -march | Original (seconds) | Modified (seconds) | Improvement | Original (Giga-cycles) | Modified (Giga-cycles) | Improvement
| --- | --- | --- | --- | --- | --- | --- | --- |
clang | skylake | 4.916 | 4.788 | 2.62% | 134.955 | 131.069 | 2.88%
clang | x86-64 | 4.965 | 4.778 | 3.75% | 134.378 | 130.901 | 2.59%
gcc | skylake | 4.871 | 4.829 | 0.87% | 134.192 | 131.280 | 2.17%
gcc | x86-64 | 4.832 | 4.646 | 3.85% | 134.118 | 130.446 | 2.74%

Each reported value is the average from 5 runs. Each run was preceded by 30 seconds of sleep so that the CPUs can start non-throttled.