@kimwalisch 2022-01-11
> I've been trying to squeeze a bit more performance of out the multi-buffer PreSieve idea, and at the same time reduce the total memory taken by these buffers. Let me know what you think.

I think it is great! I tested it on 2 of my CPUs and it indeed ran faster using both a single thread and multiple threads. On top of that the new code also uses less memory! I slightly improved the distribution of the primes in the bufferPrimes vector, this saves about 100 bytes.

Note that last weekend I tested whether a single shared PreSieve object (i.e. a singleton using primes <= 59) would improve performance. But it ran equally fast. I prefer the current one PreSieve object per thread design as this should theoretically scale better on servers with a large number of CPU cores (because many threads accessing the same memory could cause a bottleneck).